chore(0.11.11): 96e-T1 — code-shape KEEP 8 class relocate 進 frontend_contracts/（純搬移零行為變更）
- 新檔 tests/unit/frontend_contracts/test_contract_code_shape.py（977 行 / 8 class / 54 test）
- CoverState/SearchAllRace/LightboxAnimation/LightboxStateFirst/ShowcaseReactiveScope/
  ExternalManagerSwitchMode/NavigateLoadMore/NextLightboxLoadMore（CD-96-10 KEEP-pytest-justified）
- pure-move gate（CD-96e-2）：collect test-id 差集空（54==54）+ byte-for-byte 驗證
  （唯二允許差異：lint-guard 標籤註解 + NavigateLoadMore inline NAVIGATION_JS 深度+1）
- 源檔 928 行純刪除（0 added）、5 個 module 常數複製非剪走、slim-residual 兩 class 未動
- test_frontend_lint.py：6,866→5,938 行 / 81→73 class
